#3046 优化代码
