+++ src/phar/src/Ast/Ast.php
@@ -11,7 +11,6 @@ declare(strict_types=1);
  */
 namespace Hyperf\Phar\Ast;
 
-use PhpParser\Node;
 use PhpParser\NodeTraverser;
 use PhpParser\Parser;
 use PhpParser\ParserFactory;
@@ -45,27 +44,6 @@ class Ast
         }
         $stmts = $this->astParser->parse($code);
         $stmts = $traverser->traverse($stmts);
-        array_push($stmts, $this->createReturn());
         return $this->printer->prettyPrintFile($stmts);
-    }
-
-    protected function createReturn(): Node\Stmt\Return_
-    {
-        $funcCall = new Node\Expr\FuncCall(new Node\Name('array_replace'));
-        $funcCall->args = [
-            new Node\Arg(new Node\Expr\Variable('result')),
-            $this->createScanArg(),
-        ];
-        return new Node\Stmt\Return_($funcCall);
-    }
-
-    protected function createScanArg(): Node\Arg
-    {
-        $array = new Node\Expr\Array_();
-        $array->items[] = new Node\Expr\ArrayItem(
-            new Node\Expr\ConstFetch(new Node\Name('true')),
-            new Node\Scalar\String_('scan_cacheable')
-        );
-        return new Node\Arg($array);
     }
 }
